FWA-27: Fix bg scroll on iOS safari
In a prior attempt, I disabled the propagation of the touchmove event
from the pull up menu to the body (which was scrolling, when scroll
occurred on the pull up). While I confirmed the event does NOT
propagate, the browser still scrolls the content. By calling
preventDefault() on the event (effectively disabling scrolling for the
pull up menu also) the scrolling is disabled.