FWA-43: Fix login requiring a manual close
The code stopped listening for auth0 events after an error occurred. If
a user typed in the wrong password, we would stop listening, and
our event listener was responsible for closing the form.

The code now listens until either the form is manually closed, there is
success, or an unrecoverable error happens. At which point, we close the
form.