FWA-51: Prefer a non-overlay approach to the pull up menu
Previously, when the user clicked "Passes" from the initial page on
mobile, we would render an overlay. This overlay could have scrolling
content in the case where there were many tickets to choose from. At the
limits of scrolling this ticket list, the scroll would begin scrolling
the parent page. The only way to prevent that scroll was to do some js
hacking: cancelling specific scroll events on the ticket view. This is because
Safari transfers the scroll to the parent irrespective of the rules of
event capturing and event propagation.

Now, there is no overlay, so there is no risk of background scolling.
The entire page is the ticket menu or it is the event page.