Update scaffolding with mailchimp form submission
In this updated approach we submit a form to mailchimp for the user to
subscribe. The response opens up in a new tab/window. If the user
provided an invalid email, the mailchimp form will provide validation.

We forward along the user's first and last name if we have it otherwise
we pass along the empty string. It is unlikely that we will have it (as
a user would have to login, without being prompted to in the first
place), but it is optional.